requirements.txt
pydub~=0.25.1
simpleaudio~=1.0.4
adafruit-circuitpython-neopixel~=6.2.2
adafruit-blinka~=6.15.0